Alhamdulillah! routes validations added -  users, vehicles

+++ src/modules/vehicles/vehicles.controllers.ts
@@ -115,17 +115,27 @@ const updateVehicle = async (req: Request, res: Response) => {
 
 const deleteVehicle = async (req: Request, res: Response) => {
   try {
-    const result = await vehicleServices.deleteVehicle(
+    const existing = await vehicleServices.getSingleVehicle(
       req.params.vehicleId as string
     );
 
+    if (existing.rows[0]?.availability_status === "booked") {
+      return res.status(404).json({
+        success: false,
+        message: "Vehicle cannot be deleted because it is currently booked.",
+      });
+    }
+
+    const result = await vehicleServices.deleteVehicle(
+      req.params.vehicleId as string
+    );
     if (result.rowCount === 0) {
-      res.status(404).json({
+      return res.status(404).json({
         success: false,
         message: "Vehicle not found",
       });
     } else {
-      res.status(200).json({
+      return res.status(200).json({
         success: true,
         message: "Vehicle deleted successfully",
       });
